LayoutTests/platform/ios-simulator/editing/inserting/paragraph-outside-nested-divs-expected.txt

CONSOLE MESSAGE: line 14: wrong nestedDiv? <br>
When inserting a new line, we should break out of nested divs.

FAIL

first
second

This should be in nested divs.third
This should be in a single div.
